Move the image checks into a ImageCheck class
