+++ lib/envato_optimiser.rb
@@ -1,5 +1,7 @@
 require 'envato_optimiser/version'
 
 module EnvatoOptimiser
-  autoload :Item, 'envato_optimiser/item'
+  autoload :Item,       'envato_optimiser/item'
+  autoload :Check,      'envato_optimiser/check'
+  autoload :ImageCheck, 'envato_optimiser/checks/image_check'
 end
