+++ lib/envato_optimiser/check.rb
@@ -1,0 +1,4 @@
+module EnvatoOptimiser
+  class Check
+  end
+end
